requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0

# HTTP 클라이언트
httpx==0.26.0

# AI/ML 라이브러리
openai==1.40.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-qdrant==0.1.1
langchain-text-splitters==0.2.1
langgraph==0.2.17

# 벡터 데이터베이스
qdrant-client==1.9.0

# BM25 검색
rank_bm25==0.2.2

# 환경 변수 및 설정
python-dotenv==1.0.0
pydantic==2.4.2

# Google AI (Gemini)
google-generativeai==0.8.3

# Anthropic (Claude)
anthropic==0.34.0

# 웹 크롤링 라이브러리
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 테스트 라이브러리
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

python-dateutil
